Make remove button compatible with gtm4wp
Added the .remove class to the button to work with the js that gtm4wp is using. as the class is also present in the storefront theme i assume that could add compatibility to other plugins as well.

Ovewrites to some woocommerce css classes had to be done to keep the original styling